Move tests / move bits of readme
