+++ fronts-client/src/actions/__tests__/frontscards.spec.ts
@@ -1,9 +1,9 @@
 import configureMockStore from 'redux-mock-store';
 import fetchMock from 'fetch-mock';
 import thunk from 'redux-thunk';
-import { createArticleEntitiesFromDrop, cardsReceived } from '../Cards';
+import { createArticleEntitiesFromDrop, cardsReceived } from '../../shared/actions/Cards';
 import initialState from 'fixtures/initialState';
-import { capiArticle } from '../../../fixtures/shared';
+import { capiArticle } from '../../fixtures/shared';
 import { actionNames as externalArticleActionNames } from 'bundles/externalArticlesBundle';
 import { createCard } from 'util/card';
 import { RefDrop } from 'util/collectionUtils';
